third_party/blink/web_tests/platform/linux/virtual/webnn-service-with-gpu/external/wpt/webnn/conformance_tests/arg_min_max.https.any_gpu-expected.txt

This is a testharness.js-based test.
Found 12 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] argMin uint32 4D tensor, axis=1, all options
  promise_test: Unhandled rejection with value: object "TypeError: Unsupported data type, input 'input' data type uint32 must be one of [float32,float16,int32,int8,uint8]."
[FAIL] argMin int64 4D tensor, axis=0, all options
  promise_test: Unhandled rejection with value: object "TypeError: Unsupported data type, input 'input' data type int64 must be one of [float32,float16,int32,int8,uint8]."
[FAIL] argMin uint64 4D tensor, axis=1, all options
  promise_test: Unhandled rejection with value: object "TypeError: Unsupported data type, input 'input' data type uint64 must be one of [float32,float16,int32,int8,uint8]."
[FAIL] argMax float32 4D tensor, axis=3, options.keepDimensions=true
  assert_less_than_equal: assert_array_approx_equals_ulp: test argMax int32 actual 3 should be close enough to expected 1 by ULP distance: expected a number less than or equal to 0 but got 2
[FAIL] argMax float32 4D tensor, axis=3, options.keepDimensions=false
  assert_less_than_equal: assert_array_approx_equals_ulp: test argMax int32 actual 3 should be close enough to expected 1 by ULP distance: expected a number less than or equal to 0 but got 2
[FAIL] argMax float32 4D tensor, axis=3, explicit options.outputDataType='int32'
  assert_less_than_equal: assert_array_approx_equals_ulp: test argMax int32 actual 3 should be close enough to expected 1 by ULP distance: expected a number less than or equal to 0 but got 2
[FAIL] argMax float16 4D tensor, axis=3, options.keepDimensions=true
  assert_less_than_equal: assert_array_approx_equals_ulp: test argMax int32 actual 3 should be close enough to expected 1 by ULP distance: expected a number less than or equal to 0 but got 2
[FAIL] argMax float16 4D tensor, axis=3, options.keepDimensions=false
  assert_less_than_equal: assert_array_approx_equals_ulp: test argMax int32 actual 3 should be close enough to expected 1 by ULP distance: expected a number less than or equal to 0 but got 2
[FAIL] argMax float16 4D tensor, axis=3, explicit options.outputDataType='int32'
  assert_less_than_equal: assert_array_approx_equals_ulp: test argMax int32 actual 3 should be close enough to expected 1 by ULP distance: expected a number less than or equal to 0 but got 2
[FAIL] argMax uint32 4D tensor, axis=1, all options
  promise_test: Unhandled rejection with value: object "TypeError: Unsupported data type, input 'input' data type uint32 must be one of [float32,float16,int32,int8,uint8]."
[FAIL] argMax int64 4D tensor, axis=0, all options
  promise_test: Unhandled rejection with value: object "TypeError: Unsupported data type, input 'input' data type int64 must be one of [float32,float16,int32,int8,uint8]."
[FAIL] argMax uint64 4D tensor, axis=1, all options
  promise_test: Unhandled rejection with value: object "TypeError: Unsupported data type, input 'input' data type uint64 must be one of [float32,float16,int32,int8,uint8]."
Harness: the test ran to completion.
